requirements.txt
Django==1.10.2
MySQL-python==1.2.5
Pygments==2.1.3
djangorestframework==3.4.7
